Track archived and fork in Repository record
This is primarily useful to inspect the values in GET (or PUT)
responses, for example when we build import and we'd like to skip based
on it. `archived` can also be used to archive a repository, but `fork`
will always be ignored. There is a different API for dealing with forks,
which we don't have plans to implement yet.