Default more values in the actual codecs
If GitHub has a default itself, we can make the field optional with that
default and then users need not specify it in their yaml at all. This
slimmed down our defaults.yaml quite a bit.

We can also do archived and fork this way, so those fields (which we
added for our own read-only usecases) don't have to show up to users if
they don't care.